+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [STVP](https://drive.mia.kiwi/index.php/s/PLGxSTNDxSTVP).
 
+## [25.0.2] - 2025-10-18
+
+### Added
+
+- Fields track what form they're a member of (`form` attribute`)
+- Field validator class`FieldValidator`, executed after HTML5 constraints
+- Special validation options for `PasswordField`: `includeNumbers`, `includeSymbols`, `includeUppercase`, `includeLowercase`, and `min`
+- More details and example in README
+
+### Changed
+
+- Casing in comments for consistency
+
+
+
 ## [25.0.1] - 2025-10-18
 
 ### Added